+++ source/Nuke.Common/Tools/OpenCover/OpenCoverTasks.cs
@@ -4,7 +4,6 @@
 
 using System;
 using System.Linq;
-using JetBrains.Annotations;
 using Nuke.Core;
 using Nuke.Core.Tooling;
 
@@ -31,18 +30,5 @@ namespace Nuke.Common.Tools.OpenCover
                 "*/*.Designer.cs",
                 "*/*.g.cs",
                 "*/*.g.i.cs");
-
-        [CanBeNull]
-        private static IProcess StartProcess(OpenCoverSettings toolSettings, ProcessSettings processSettings = null)
-        {
-            var testAction = toolSettings.TestAction.NotNull("testAction != null");
-            var capturedStartInfo = ProcessTasks.CaptureProcessStartInfo(testAction);
-            toolSettings = toolSettings
-                .SetTargetPath(capturedStartInfo.ToolPath)
-                .SetTargetArguments(capturedStartInfo.Arguments)
-                .SetTargetDirectory(capturedStartInfo.WorkingDirectory);
-
-            return ProcessTasks.StartProcess(toolSettings, processSettings);
-        }
     }
 }